% Generated by roxygen2 (4.0.2): do not edit by hand
\name{celcius_to_kelvin}
\alias{celcius_to_kelvin}
\title{Convert degree Celsius to degree Kelvin}
\usage{
celcius_to_kelvin(value)
}
\arguments{
\item{value}{Temperature (in degree Celsius) value to convert to degree Kelvin}
}
\value{
Temperature value in degree Kelvin
}
\description{
\code{celsius_to_kelvin} converts a temperature value in degree Celsius to degree Kelvin.
}
\seealso{
\code{\link{kelvin_to_celsius}}
}

