% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/darcy_to_square_metre.R
\name{darcy_to_square_metre}
\alias{darcy_to_square_metre}
\title{Convert darcy units to square metre}
\usage{
darcy_to_square_metre(value)
}
\arguments{
\item{value}{Permeability (in darcy units) value to convert to square metre}
}
\value{
Permeability value in square metre
}
\description{
\code{darcy_to_square_metre} converts a permeability value in darcy units to square metre.
}
\seealso{
\code{\link{square_metre_to_darcy}}
}

