% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_celcius.R
\name{fahrenheit_to_celcius}
\alias{fahrenheit_to_celcius}
\title{Convert degree Fahrenheit to degree Celsius}
\usage{
fahrenheit_to_celcius(value)
}
\arguments{
\item{value}{Temperature (in degree Fahrenheit) value to convert to degree Celsius}
}
\value{
Temperature value in degree Celsius
}
\description{
\code{kelvin_to_celsius} converts a temperature value in degree Fahrenheit to degree Celsius.
}
\seealso{
\code{\link{celsius_to_kelvin}}, \code{\link{celsius_to_fahrenheit}}, \code{\link{kelvin_to_celsius}}, \code{\link{kelvin_to_fahrenheit}}, \code{\link{fahrenheit_to_kelvin}}
}

