% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_kelvin.R
\name{fahrenheit_to_kelvin}
\alias{fahrenheit_to_kelvin}
\title{Convert degree Fahrenheit to degree Kelvin}
\usage{
fahrenheit_to_kelvin(value)
}
\arguments{
\item{value}{Temperature (in degree Fahrenheit) value to convert to degree Kelvin}
}
\value{
Temperature value in degree Kelvin
}
\description{
\code{fahrenheit_to_kelvin} converts a temperature value in degree Fahrenheit to degree Kelvin.
}
\seealso{
\code{\link{celsius_to_kelvin}}, \code{\link{celsius_to_fahrenheit}}, \code{\link{kelvin_to_celsius}}, \code{\link{kelvin_to_fahrenheit}}, \code{\link{fahrenheit_to_celsius}}
}

