% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hydraulic_conductivity_to_permeability}
\alias{hydraulic_conductivity_to_permeability}
\title{Convert hydraulic conductivity to permeability}
\usage{
hydraulic_conductivity_to_permeability(K, mu = 0.001308, rho = 999.7,
  g = 9.80665, t = 10)
}
\arguments{
\item{K}{Hydraulic conductivity of the medium (m/s)}

\item{mu}{Dynamic viscosity of the fluid; defaults to 1.308E-3 Pa*s}

\item{rho}{Density of the fluid; defaults to 999.7 kg/m³}

\item{g}{Gravitational acceleration; defaults to 9.80665 m/s²}

\item{t}{Temperature; defaults to 10°C}
}
\value{
Intrinsic permeability value
}
\description{
Converts a hydraulic conductivity value to a value for intrinsic permeability.
}
\details{
\code{mu} and \code{rho} are automatically determined in function of t in a temperature range between 0 and 100 °C, unless the values are explicitely specified.
}
\seealso{
\code{\link{permeability_to_hydraulic_conductivity}}
}

