% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inch_to_metre.R
\name{inch_to_metre}
\alias{inch_to_metre}
\title{Convert inch to metre}
\usage{
inch_to_metre(value)
}
\arguments{
\item{value}{Length (in inches) to convert to metres}
}
\value{
Length in metres
}
\description{
\code{inch_to_metre} converts a length in inches to metres.
}

