% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelvin_to_fahrenheit.R
\name{kelvin_to_fahrenheit}
\alias{kelvin_to_fahrenheit}
\title{Convert degree Kelvin to degree Fahrenheit}
\usage{
kelvin_to_fahrenheit(value)
}
\arguments{
\item{value}{Temperature (in degree Kelvin) value to convert to degree Fahrenheit}
}
\value{
Temperature value in degree Fahrenheit
}
\description{
\code{kelvin_to_celsius} converts a temperature value in degree Kelvin to degree Fahrenheit.
}
\seealso{
\code{\link{celsius_to_kelvin}}, \code{\link{celsius_to_fahrenheit}}, \code{\link{kelvin_to_celsius}}, \code{\link{fahrenheit_to_celsius}}, \code{\link{fahrenheit_to_kelvin}}
}

