% Generated by roxygen2 (4.0.2): do not edit by hand
\name{permeability_to_hydraulic_conductivity}
\alias{permeability_to_hydraulic_conductivity}
\title{Convert permeability to hydraulic conductivity}
\usage{
permeability_to_hydraulic_conductivity(k, mu = 0.001308, rho = 999.7,
  g = 9.80665, t = 10)
}
\arguments{
\item{k}{Intrinsic permeability of the medium (in square metres)}

\item{mu}{Dynamic viscosity of the fluid; defaults to 1.308E-3 Pa*s}

\item{rho}{Density of the fluid; defaults to 999.7 kg/m³}

\item{g}{Gravitational acceleration; defaults to 9.80665 m/s²}

\item{t}{Temperature; defaults to 10°C}
}
\value{
Hydraulic conductivity value
}
\description{
Converts an intrinsic permeability value to a value for hydraulic conductivity.
}
\details{
\code{mu} and \code{rho} are automatically determined in function of t in a temperature range between 0 and 100 °C, unless the values are explicitely specified.
}
\seealso{
\code{\link{hydraulic_conductivity_to_permeability}}
}

