% Generated by roxygen2 (4.0.2): do not edit by hand
\name{point_dilution_method}
\alias{point_dilution_method}
\title{Point dilution method}
\usage{
point_dilution_method(conc, time, conc_init, conc_final, time_init, time_final,
  well_diam, well_head, filter_length, hose_diam = 0.0127, hose_length = 10)
}
\arguments{
\item{conc}{Concentration vector}

\item{time}{Time vector}

\item{conc_init}{Initial concentration}

\item{conc_final}{Final concentration (background)}

\item{time_init}{Starting time of the point dilution test}

\item{time_final}{Ending time of the point dilution test}

\item{well_diam}{Well diameter}

\item{well_head}{Water column height inside the well}

\item{filter_length}{Length of the filter}

\item{hose_diam}{Diameter of the recirculation hose}

\item{hose_length}{Length of the hose outside of the water column}
}
\value{
Apparent specific discharge (length/time)
}
\description{
Provides the apparent specific discharge from a dilution curve.
}
\details{
To obtain the specific discharge, the apparent specific discharge should be divided by the flow distortion factor.
}
\seealso{
\code{\link{well_flow_distortion_factor}}
}

