% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/square_metre_to_darcy.R
\name{square_metre_to_darcy}
\alias{square_metre_to_darcy}
\title{Convert square metre to darcy units}
\usage{
square_metre_to_darcy(value)
}
\arguments{
\item{value}{Permeability (in square metre) value to convert to darcy units}
}
\value{
Permeability value in darcy units
}
\description{
\code{square_metre_to_darcy} converts a permeability value in square metre to darcy units.
}
\seealso{
\code{\link{darcy_to_square_metre}}
}

