% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tinyperm_to_air_permeability.R
\name{tinyperm_to_air_permeability}
\alias{tinyperm_to_air_permeability}
\title{Tinyperm value to air permeability conversion}
\usage{
tinyperm_to_air_permeability(x, calibration = "Huysmans2008")
}
\arguments{
\item{x}{TinyPerm II values}

\item{calibration}{Which calibration to use: Huysmans2008 or factory}
}
\value{
Air permeability (m²)
}
\description{
Converts a value given by the TinyPerm II air permeameter into a value for air permeability
}
\seealso{
\code{\link{tinyperm_to_hydraulic_conductivity}}, \code{\link{air_permeability_to_hydraulic_conductivity}}
}

