% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tinyperm_to_hydraulic_conductivity}
\alias{tinyperm_to_hydraulic_conductivity}
\title{Tinyperm value to hydraulic conductivity conversion}
\usage{
tinyperm_to_hydraulic_conductivity(x, calibration = c("Huysmans2008",
  "Iversen2003"))
}
\arguments{
\item{x}{TinyPerm II values}

\item{calibration}{Which calibrations to use: see details.}
}
\value{
Hydraulic conductivity (m/s)
}
\description{
Converts a value given by the TinyPerm II air permeameter into a value for hydraulic conductivity
}
\details{
\code{calibration} is a vector of two strings, with the first element Huysmans2008 or factory, and the second element Iversen2003 or Loll1999
}
\seealso{
\code{\link{tinyperm_to_air_permeability}}, \code{\link{air_permeability_to_hydraulic_conductivity}}
}

