% Generated by roxygen2 (4.0.2): do not edit by hand
\name{well_flow_distortion_factor}
\alias{well_flow_distortion_factor}
\title{Well flow distortion factor}
\usage{
well_flow_distortion_factor(r_s, r_f, r_a, K_s, K_f, K_a, equation = "Halevy",
  recursion_level = 100)
}
\arguments{
\item{equation}{'Halevy', 'Ogilvi', or 'Bidaux and Tsang'}

\item{param}{explanation}
}
\value{
Returned value
}
\description{
\code{function} description
}
\details{
More detailed explanation
}
\seealso{
\code{\link{related_function}}
}

