% Generated by roxygen2 (4.0.2): do not edit by hand
\name{well_screen_hydraulic_conductivity}
\alias{well_screen_hydraulic_conductivity}
\title{Well screen hydraulic conductivity}
\usage{
well_screen_hydraulic_conductivity(r_s, r_f, slot_area, slot_perimeter,
  slot_screen_area_ratio, slot_resistance_coef = 12, mu = 0.001308,
  rho = 999.7, g = 9.80665, t = 10)
}
\arguments{
\item{param}{explanation}
}
\value{
Returned value
}
\description{
\code{function} description
}
\details{
More detailed explanation
}
\seealso{
\code{\link{related_function}}
}

