% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/well_screen_resistance.R
\name{well_screen_resistance}
\alias{well_screen_resistance}
\title{Well screen resistance}
\usage{
well_screen_resistance(r_s, r_f, slot_area, slot_perimeter)
}
\arguments{
\item{r_s}{inner diameter of well screen (mm)}

\item{r_f}{outer diameter of well screen (mm)}

\item{slot_area}{area of the slot (mm²)}

\item{slot_perimeter}{(mm)}
}
\value{
Returned value
}
\description{
\code{well_screen_resistance} provides the
}
\details{
More detailed explanation
}
\references{
Drost et al. (1968)
Palmer (1993)
}
\seealso{
\code{\link{well_screen_hydraulic_conductivity}}
\code{\link{well_flow_distortion_factor}}
}

