% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius_to_kelvin.R
\name{celcius_to_kelvin}
\alias{celcius_to_kelvin}
\title{Convert degree Celsius to degree Kelvin}
\usage{
celcius_to_kelvin(value)
}
\arguments{
\item{value}{Temperature (in degree Celsius) value to convert to degree Kelvin}
}
\value{
Temperature value in degree Kelvin
}
\description{
\code{celsius_to_kelvin} converts a temperature value in degree Celsius to degree Kelvin.
}
\seealso{
\code{\link{celsius_to_fahrenheit}}, \code{\link{kelvin_to_celsius}}, \code{\link{kelvin_to_fahrenheit}}, \code{\link{fahrenheit_to_celsius}}, \code{\link{fahrenheit_to_kelvin}}
}

