% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kelvin_to_celsius.R
\name{kelvin_to_celsius}
\alias{kelvin_to_celsius}
\title{Convert degree Kelvin to degree Celsius}
\usage{
kelvin_to_celsius(value)
}
\arguments{
\item{value}{Temperature (in degree Kelvin) value to convert to degree Celsius}
}
\value{
Temperature value in degree Celsius
}
\description{
\code{kelvin_to_celsius} converts a temperature value in degree Kelvin to degree Celsius.
}
\seealso{
\code{\link{celsius_to_kelvin}}, \code{\link{celsius_to_fahrenheit}}, \code{\link{kelvin_to_fahrenheit}}, \code{\link{fahrenheit_to_celsius}}, \code{\link{fahrenheit_to_kelvin}}
}

